add clearconsole.class
