+++ src/main/java/com/infoshareacademy/App.java
@@ -1,6 +1,5 @@
 package com.infoshareacademy;
 
-import com.infoshareacademy.menu.MenuLogin;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -9,21 +8,6 @@ import org.slf4j.LoggerFactory;
  */
 public class App {
     private static final Logger STDOUT = LoggerFactory.getLogger("CONSOLE_OUT");
-
-    public static void clearConsole(){
-        try{
-            final String os = System.getProperty("os.name");
-            if (os.contains("Windows")){
-                Runtime.getRuntime().exec("cls");
-            }
-            else {
-                Runtime.getRuntime().exec("clear");
-            }
-        }
-        catch (final Exception e){
-
-        }
-    }
 
     public static void main(String[] args) {
         STDOUT.info("jAvalanche \n");
@@ -34,9 +18,9 @@ public class App {
             STDOUT.info(temp);
         }
         STDOUT.info(holidaysJsonData.toString());
-        
-        MenuLogin menuLogin = new MenuLogin();
-        menuLogin.Login();
+
+//        MenuLogin menuLogin = new MenuLogin();
+//        menuLogin.Login();
 
     }
 }
