+++ src/main/java/com/infoshareacademy/menu/ClearConsole.java
@@ -1,0 +1,18 @@
+package com.infoshareacademy.menu;
+
+public class ClearConsole {
+    public static void clearConsole(){
+        try{
+            final String os = System.getProperty("os.name");
+            if (os.contains("Windows")){
+                Runtime.getRuntime().exec("cls");
+            }
+            else {
+                Runtime.getRuntime().exec("clear");
+            }
+        }
+        catch (final Exception e){
+
+        }
+    }
+}
