+++ src/main/java/com/infoshareacademy/menu/MenuSearch.java
@@ -1,18 +1,15 @@
 package com.infoshareacademy.menu;
 
-import com.infoshareacademy.App;
-import com.infoshareacademy.menu.MainMenu;
-
 import java.util.Scanner;
 
 public class MenuSearch {
 
     public void menuSearch(){
-        App.clearConsole();
+        ClearConsole.clearConsole();
         Scanner scanner = new Scanner(System.in);
         MainMenu.STDOUT.info("Choose option");
         MainMenu.STDOUT.info("1. Searching holidays");
-        MainMenu.STDOUT.info("2. Searching holidays by employer");
+        MainMenu.STDOUT.info("2. Searching holidays by employee");
         MainMenu.STDOUT.info("3. Searching holidays by team");
         int option = scanner.nextInt();
         switch (option){
@@ -35,15 +32,15 @@ public class MenuSearch {
     }
 
     public void searchingHolidays(){
-        App.clearConsole();
+        ClearConsole.clearConsole();
         MainMenu.STDOUT.info("Searching holidays");
     }
     public void searchingByEmployee(){
-        App.clearConsole();
+        ClearConsole.clearConsole();
         MainMenu.STDOUT.info("Searching holidays by employer");
     }
     public void searchingByTeam(){
-        App.clearConsole();
+        ClearConsole.clearConsole();
         MainMenu.STDOUT.info("Searching holidays by team");
     }
 }
